Add project-specific runtimes and much else (desc)
Control Panel now contains a "Problems" section so users don't have to hunt through the console to see errors, and errors now contain an optional `solution` property to helpfully explain what they should do.

In addition, preferences are starting to move to the control panel.

Finally, runtimes which are too old or too new for Constructor to use at the moment are excluded.