refactor: partially rewrite job error system
Rewrote the Job error parsing system to use regexes. This makes it a bit easier to work with and improve (and say goodbye to the "spaghetti code" error message :P)

I might go through later on with a second pass to parse errors *more* nicely in future, such as parsing script names to link to them in code. If I recall, Builder figured out how to do that, so may look there for a reference of how to call into Ace for the same behaviour.