deps: add `scope-extensions-js` library
I'm not generally one to bring in libraries but this one was nice and small, and lets me do some stuff much more cleanly. With modifications to bring to regular JS, plus some fixes on signatures with `undefined`, and removal of redundant `apply` since it requires longer `function()` over arrows, and `this` is longer than `it` anyway.