refactor: death to the singletons, pick runtime for project automatically!
Another extremely large commit. This work encompasses moving all of the various plugin event listeners into the central `GMConstructor` instance so that we have proper fine-grained control over dependencies for various aspects of the plugin, especially regarding what is done for projects which Constructor is not compatible with (<2.3)

Additionally the "global" version dropdown has been removed, in favour of projects either:

a) allowing Constructor to pick a runtime implicitly, based on its knowledge about compatibility, or
b) picking a runtime manually from the installed list.