style: cleaner syntax for hamburger menu calling current plugin inst
The hamburger menu is a bit silly in that we have to give over control to Electron's menu system, and we cannot remove items we've already created previously. The best we can do is write their click methods to be separate from the plugin itself and call the current plugin instance, if it exists, otherwise do nothing.