fix: hamburger menu items enabled/disabled at the wrong times
This has also removed the very last instance of a non-central call to `GMEdit.on/off`, which was actually the cause of the problem, as the hamburger menu "singleton" file was registering its event listeners sooner than the plugin itself, and would thus get the wrong state.

Additionally there's no more `window.ConstructorPlugin`-based discovery, as since we *can* clear a *sub*menu, we just do that and re-add new items which use the up-to-date click callbacks.