feat: tabbed bottom panel output!
not finished by any stretch, but you can now display the logs in a separate bottom pane, similar to the GM IDE. this required some refactoring to how the output log works, and that now exists as a "client" which has a "display server", of which the server may be either the standard tabbed display, or a tab in this new bottom pane.

currently you can only use the new mode, as the preference is not exposed for it. css needs a lot of work too. bottom pane is extremely disgusting internally as its a standin for something that'll hopefully make it into gmedit later on properly.